+++ packages/store/src/persistence/blob/memory-storage.ts
